+++ legacy.js
@@ -2,7 +2,6 @@ module.exports = {
 	'extends': [
 		'eslint-config-smartthings/rules/best-practices',
 		'eslint-config-smartthings/rules/errors',
-		'eslint-config-smartthings/rules/legacy',
 		'eslint-config-smartthings/rules/node',
 		'eslint-config-smartthings/rules/strict',
 		'eslint-config-smartthings/rules/style',