+++ rules/errors.js
@@ -1,7 +1,5 @@
 module.exports = {
 	'rules': {
-		// disallow trailing commas in object literals
-		'comma-dangle': [2, 'never'],
 		// disallow assignment in conditional expressions
 		'no-cond-assign': [2, 'always'],
 		// disallow use of console
@@ -48,6 +46,8 @@ module.exports = {
 		'no-regex-spaces': 2,
 		// disallow sparse arrays
 		'no-sparse-arrays': 2,
+		// Avoid code that looks like two expressions but is actually one
+		'no-unexpected-multiline': 1,
 		// disallow unreachable statements after a return, throw, continue, or break statement
 		'no-unreachable': 2,
 		// disallow control flow statements in finally blocks
@@ -57,8 +57,6 @@ module.exports = {
 		// ensure JSDoc comments are valid
 		'valid-jsdoc': 0,
 		// ensure that the results of typeof are compared against a valid string
-		'valid-typeof': 2,
-		// Avoid code that looks like two expressions but is actually one
-		'no-unexpected-multiline': 1
+		'valid-typeof': 2
 	}
 };